# Log warning about "function" meter re-registration