>I also realized that while we added logging for gauge double registration, any of the "functional" meters are affected - FunctionCounter, FunctionTimer. That's relevant to KafkaMetrics because it only registers Gauges and FunctionCounters. /cc @jonatan-ivanov

found by @shakuzen, see: https://github.com/micrometer-metrics/micrometer/issues/5757#issuecomment-2760837057

See gh-5616
See gh-5688
See gh-5617